配置mapstruct，添加cors配置

+++ huiyu-service/service-core/src/main/java/com/huiyu/service/core/controller/ModelController.java
@@ -55,7 +55,7 @@ public class ModelController {
      * 批量删除
      */
     @PostMapping("/removeByIds")
-    public R<?> remove(@RequestBody List<Integer> ids) {
+    public R<?> remove(@RequestBody List<Long> ids) {
         return R.ok(modelService.removeByIds(ids));
     }
 }
